Add operations script with update command
This can now be run from inside a project. Best way to do this, is to symlink these files in the project _operations folder (or copy if you want to customize them).

Requires config.sh with local variables to be present in _operations. This file will be generated by Romanesco Seed for new installations.